refactor: extract `argus-providers` crate for EVM RPC data source
