+++ Cargo.toml
@@ -8,6 +8,7 @@ resolver = "2"
 [workspace.dependencies]
 alloy = { version = "1.0.37", features = ["essentials"] }
 argus-core = { path = "crates/argus-core" }
+argus-providers = { path = "crates/argus-providers" }
 argus-store = { path = "crates/argus-store" }
 omnihook = { path = "crates/omnihook" }
 rhai-analyzer = { path = "crates/rhai-analyzer" }
@@ -68,6 +69,7 @@ edition = "2024"
 [dependencies]
 alloy.workspace = true
 argus-core.workspace = true
+argus-providers.workspace = true
 argus-store.workspace = true
 arc-swap.workspace = true
 async-nats.workspace = true
