+++ crates/argus-providers/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "argus-providers"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+argus-core.workspace = true
+alloy.workspace = true
+arc-swap.workspace = true
+async-trait.workspace = true
+futures.workspace = true
+tower.workspace = true
+tracing.workspace = true
+thiserror.workspace = true
