+++ crates/argus-providers/src/lib.rs
@@ -1,0 +1,6 @@
+//! EVM RPC provider implementation for Argus.
+
+pub mod block_fetcher;
+pub mod rpc;
+
+pub use rpc::{EvmRpcSource, ProviderError, create_provider};
