+++ src/engine/block_ingestor.rs
@@ -12,7 +12,8 @@ use argus_core::{
 use tokio::sync::mpsc;
 use tokio_util::sync::CancellationToken;
 
-use crate::{context::AppMetrics, engine::filtering::FilteringEngine, providers::block_fetcher};
+use crate::{context::AppMetrics, engine::filtering::FilteringEngine};
+use argus_providers::block_fetcher;
 
 /// The BlockIngestor service.
 ///
