+++ src/lib.rs
@@ -14,6 +14,5 @@ pub mod http_client;
 pub mod http_server;
 pub mod loader;
 pub mod monitor;
-pub mod providers;
 pub mod supervisor;
 pub mod test_helpers;
