--- src/providers/mod.rs
@@ -1,4 +1,0 @@
-//! Provider modules for various data sources.
-
-pub mod block_fetcher;
-pub mod rpc;
