+++ src/supervisor/builder.rs
@@ -4,10 +4,8 @@
 use argus_core::persistence::traits::{AppRepository, KeyValueStore};
 
 use super::Supervisor;
-use crate::{
-    context::{AppContext, AppMetrics},
-    providers::rpc::EvmRpcSource,
-};
+use argus_providers::EvmRpcSource;
+use crate::context::{AppContext, AppMetrics};
 
 /// Marker for a provided component in the builder.
 pub struct Provided<T>(T);
@@ -64,7 +62,7 @@ impl<T: AppRepository + KeyValueStore + 'static>
 
         tracing::debug!(rpc_urls = ?context.config.rpc_urls, "Initializing EVM data source...");
         let evm_data_source =
-            EvmRpcSource::new(context.provider.clone(), context.monitor_manager.clone());
+            EvmRpcSource::new(context.provider.clone(), context.monitor_manager.interest_registry());
         tracing::info!(retry_policy = ?context.config.rpc_retry_config, "EVM data source initialized with fallback and retry policy.");
 
         // Construct the Supervisor with all its components from the context.
